Generalize DownloadFile and move language specific functions to language package
